Add `TInterp_t` argument to funcs to resolve interpreter instance
Functions like Declare compile code into the top-most interpreter in the stack.
If the user wants to compile the code to some other interpreter instance,
user can use the `TInterp_t` argument to specify the interpreter.

This is also required to lock the correct interpreter instance.
If `TInterp_t` is null, we assume the correct interpreter to lock is the top of the stack.
This assumption might be wrong at some cases.