[ci] Lock numba to version 0.61.2
The recent release of numba 0.62 on September 18th broke our CI with a crash in some tests. As part of supporting lates numba, we require new suppressions due to leaks in llvm::ScalarEvolution in a pass by numba through llvmlite
We should hence lock the version to 0.61 until we can fully support latest numba.